+++ agaein_web/src/components/pages/article/articleDetail/ArticleDetail.tsx
@@ -1,45 +1,28 @@
+import penguin from 'assets/image/penguin.png';
+import BookMark from 'components/molecules/BookMark';
+import Button from 'components/molecules/Button';
 import Chip from 'components/molecules/Chip';
 import Font from 'components/molecules/Font';
 import ImageCarousel from 'components/molecules/ImageCarousel/ImageCarousel';
-import {
-    Comment as CommentType,
-    Location,
-    Maybe,
-    useGetArticleQuery,
-    useGetReportsQuery,
-} from 'graphql/generated/generated';
+import Comment from 'components/organism/Comment';
+import ReactKaKaoMap from 'components/organism/ReactKakaoMap/ReactKakaoMap';
+import WitnessModal from 'components/organism/WitnessModal/WitnessModal';
+import { Comment as CommentType, useGetArticleQuery } from 'graphql/generated/generated';
+import useBookmark from 'hooks/useBookmark';
+import { Fragment, useState } from 'react';
 import { RouteComponentProps } from 'react-router';
 import { ArticleDetailParams } from 'router/params';
+import { formattedDate, YYYYMMDD } from 'utils/date';
+import { isArticle, isLFP } from 'utils/typeGuards';
 import {
-    ContainerTop,
     ArticleDetailContainer,
     ArticleDetailContentContainer,
     ArticleInfoContainer,
     ArticleMapContainer,
+    ContainerTop,
     HorizontalContainer,
     TitleAndBookMarkContainer,
 } from './ArticleDetail.style';
-import Comment from 'components/organism/Comment';
-import Button from 'components/molecules/Button';
-import BookMark from 'components/molecules/BookMark';
-import useBookmark from 'hooks/useBookmark';
-import { useState } from 'react';
-import WitnessModal from 'components/organism/WitnessModal/WitnessModal';
-import ReactKaKaoMap from 'components/organism/ReactKakaoMap/ReactKakaoMap';
-import { formattedDate, YYYYMMDD } from 'utils/date';
-import { Fragment } from 'react';
-import { isArticle, isLFP } from 'utils/typeGuards';
-import penguin from 'assets/image/penguin.png';
-
-interface WitnessArray {
-    id: string;
-    name: string;
-    address: string;
-    date: string;
-    hp?: string;
-    img?: Maybe<string>[];
-    content?: string;
-}
 
 const ArticleDetail = ({ match }: RouteComponentProps<ArticleDetailParams>) => {
     const { isBookmarked, setBookmark } = useBookmark();
@@ -52,41 +35,15 @@ const ArticleDetail = ({ match }: RouteComponentProps<ArticleDetailParams>) => {
             // TODO : 조회수 즉각반영
         },
     });
-    const reportData = useGetReportsQuery({
-        variables: {
-            articleId: match.params.id,
-        },
-    });
+
     if (loading) return <p>Loading...</p>;
     if (error) return <p>Error occur</p>;
     if (data === undefined || !isArticle(data.article)) return <p>No data</p>;
 
-    if (reportData.loading) return <p>reportData Loading...</p>;
-    if (reportData.error) return <p>reportData Error occur</p>;
     const { id, createdAt, articleDetail, view, author, comments = [], images = [] } = data.article;
 
     // ? TypeGuard로 해결할 방법을 모르겠음
     const { breed, feature, age, gender, name, location, foundDate, lostDate } = articleDetail as any;
-
-    const foundPosition: Array<Location> = [];
-    const witness: Array<WitnessArray> = [];
-    reportData.data?.reports.map((item) => {
-        let position: Location;
-        if (item !== null) {
-            position = item.location;
-            witness.push({
-                id: item.id,
-                name: item.author.nickname ?? '',
-                address: item.location.address,
-                date: item.foundDate,
-                hp: item.phoneNumber ?? '',
-                img: item.images,
-                content: String(item.content),
-            });
-            return foundPosition.push(position);
-        }
-        return null;
-    });
 
     function getTitle() {
         if (isLFP(articleDetail)) {
@@ -146,7 +103,7 @@ const ArticleDetail = ({ match }: RouteComponentProps<ArticleDetailParams>) => {
                         <ReactKaKaoMap
                             missPosition={location}
                             size={{ width: 480, height: 260 }}
-                            foundPosition={foundPosition}
+                            //foundPosition={foundPosition}
                             noClick={true}
                         />
                         <Button
@@ -165,8 +122,6 @@ const ArticleDetail = ({ match }: RouteComponentProps<ArticleDetailParams>) => {
             <WitnessModal
                 open={isOpenModal}
                 close={closeModal}
-                witness={witness}
-                foundPosition={foundPosition}
                 missPosition={location}
                 isAuthor={true}
                 articleId={id}
